Move common helper functions to OpenGLRenderer
