+++ p5/sketch/openglrenderer.py
@@ -1,7 +1,190 @@
 from abc import ABC
 import numpy as np
 
+from ..core import p5
+from ..core.constants import SType
+from ..core.primitives import Arc
+
 from vispy.gloo import Program, VertexBuffer, FrameBuffer, IndexBuffer
+from OpenGL.GLU import gluTessBeginPolygon, gluTessBeginContour, gluTessEndPolygon, gluTessEndContour, gluTessVertex
+
+
+def _tess_new_contour(vertices):
+    """Given a list of vertices, evoke gluTess to create a contour
+    """
+    gluTessBeginContour(p5.tess.tess)
+    for v in vertices:
+        gluTessVertex(p5.tess.tess, v, v)
+    gluTessEndContour(p5.tess.tess)
+
+
+def _vertices_to_render_primitive(gl_name, vertices):
+    """Returns a render primitive of gl_type with vertices in sequential order
+    """
+    return [gl_name, np.asarray(vertices), np.arange(len(vertices), dtype=np.uint32)]
+
+
+def _get_line_from_verts(vertices):
+    """Given a list of vertices, chain them sequentially in a line rendering primitive
+    """
+    n_vert = len(vertices)
+    return _get_line_from_indices(vertices, np.arange(n_vert - 1, dtype=np.uint32), np.arange(1, n_vert, dtype=np.uint32))
+
+def _get_line_from_indices(vertices, start, end):
+    """Given two columns of indices that represent edges, return a line rendering primitive
+
+    :param vertices: List of vertices
+    :type vertices: list
+
+    :param start: Array of start positions of edges in vertex indices
+    :type start: np.ndarray
+
+    :param end: Array of end positions fo edges in vertex indices
+    :type end: np.ndarray
+    """
+    start = np.asarray(start, dtype=np.uint32)
+    end = np.asarray(end, dtype=np.uint32)
+    return ['lines', np.asarray(vertices),
+            np.hstack((np.vstack(start), np.vstack(end)))]
+
+
+def _add_edges_to_primitive_list(primitive_list, vertices, start, end):
+    """Adds edges to a list of render primitives, given their start and end positions (in vertex indices)
+
+    :param start: Array of start positions of edges in vertex indices
+    :type start: np.ndarray
+
+    :param end: Array of end positions fo edges in vertex indices
+    :type end: np.ndarray
+    """
+    primitive_list.append(_get_line_from_indices(vertices, start, end))
+
+
+def _not_enough_vertices(shape, n):
+    """Returns an error string that describes how many vertices are needed
+    """
+    return "Need at least {} vertices for {}".format(n, shape)
+
+
+def _wrong_multiple(shape, n):
+    """Returns an error string that describes the # of vertices is not a multiple of n
+    """
+    return "{} requires the number of vertices to be a multiple of {}".format(shape.shape_type, n)
+
+
+def _check_shape(shape):
+    """Checks if the shape is valid using assertions
+    """
+    n_vert = len(shape.vertices)
+    if shape.shape_type in [SType.TRIANGLES, SType.TRIANGLE_FAN, SType.TRIANGLE_STRIP]:
+        assert n_vert >= 3, _not_enough_vertices(shape, 3)
+    elif shape.shape_type in [SType.LINES, SType.LINE_STRIP]:
+        assert n_vert >= 2, _not_enough_vertices(shape, 2)
+    elif shape.shape_type in [SType.QUADS, SType.QUAD_STRIP]:
+        assert n_vert >= 4, _not_enough_vertices(shape, 4)
+
+    if shape.shape_type == SType.TRIANGLES:
+        assert n_vert % 3 == 0, _wrong_multiple(shape, 3)
+    if shape.shape_type == SType.QUADS:
+        assert n_vert % 4 == 0, _wrong_multiple(shape, 4)
+
+
+def _get_borders(shape):
+    """Generates the render primitives for the borders of a given shape
+
+    :returns: ['lines', vertices, idx]
+    """
+    render_primitives = []
+    n_vert = len(shape.vertices)
+    if shape.shape_type == SType.TRIANGLES:
+        start = np.arange(n_vert)
+        end = np.arange(n_vert) + np.tile([1, 1, -2], n_vert // 3)
+        _add_edges_to_primitive_list(render_primitives, shape.vertices, start, end)
+    elif shape.shape_type == SType.TRIANGLE_STRIP:
+        start = np.concatenate((np.arange(n_vert - 1), np.arange(n_vert - 2)))
+        end = np.concatenate((np.arange(1, n_vert), np.arange(2, n_vert)))
+        _add_edges_to_primitive_list(render_primitives, shape.vertices, start, end)
+    elif shape.shape_type == SType.TRIANGLE_FAN:
+        start = np.concatenate((np.repeat([0], n_vert - 1), np.arange(1, n_vert - 1)))
+        end = np.concatenate((np.arange(1, n_vert), np.arange(2, n_vert)))
+        _add_edges_to_primitive_list(render_primitives, shape.vertices, start, end)
+    elif shape.shape_type == SType.QUADS:
+        start = np.arange(n_vert)
+        end = np.arange(n_vert) + np.tile([1, 1, 1, -3], n_vert // 4)
+        _add_edges_to_primitive_list(render_primitives, shape.vertices, start, end)
+    elif shape.shape_type == SType.QUAD_STRIP:
+        start = np.concatenate((np.arange(0, n_vert, 2), np.arange(n_vert - 2)))
+        end = np.concatenate((np.arange(1, n_vert, 2), np.arange(2, n_vert)))
+        _add_edges_to_primitive_list(render_primitives, shape.vertices, start, end)
+    elif shape.shape_type == SType.LINES:
+        start = np.arange(0, n_vert, 2)
+        end = np.arange(1, n_vert, 2)
+        _add_edges_to_primitive_list(render_primitives, shape.vertices, start, end)
+    elif shape.shape_type == SType.LINE_STRIP:
+        render_primitives.append(_get_line_from_verts(shape.vertices))
+    elif shape.shape_type == SType.TESS:
+        render_primitives.append(_get_line_from_verts(shape.vertices))
+        for contour in shape.contours:
+            render_primitives.append(_get_line_from_verts(contour))
+    return render_primitives
+
+
+def _get_meshes(shape):
+    """Generates the rendering primitives for the meshes of a given shape
+
+    :returns: [shape_type, vertices, idx]
+    """
+    render_primitives = []
+    n_vert = len(shape.vertices)
+    if shape.shape_type in [SType.TRIANGLES, SType.TRIANGLE_STRIP, SType.TRIANGLE_FAN, SType.QUAD_STRIP]:
+        gl_name = shape.shape_type.name.lower()
+        if gl_name == 'quad_strip':  # vispy does not support quad_strip
+            gl_name = 'triangle_strip'  # but it can be drawn using triangle_strip
+        render_primitives.append(_vertices_to_render_primitive(gl_name, shape.vertices))
+    elif shape.shape_type == SType.QUADS:
+        n_quad = len(shape.vertices) // 4
+        render_primitives.append(['triangles', np.asarray(shape.vertices),
+                                  np.repeat(np.arange(0, n_vert, 4, dtype=np.uint32), 6) +
+                                  np.tile(np.array([0, 1, 2, 0, 2, 3], dtype=np.uint32), n_quad)])
+    elif shape.shape_type == SType.TESS:
+        gluTessBeginPolygon(p5.tess.tess, None)
+        _tess_new_contour(shape.vertices)
+        if len(shape.contours) > 0:
+            for contour in shape.contours:
+                _tess_new_contour(contour)
+        gluTessEndPolygon(p5.tess.tess)
+        render_primitives += p5.tess.get_result()
+    return render_primitives
+
+
+def get_render_primitives(shape):
+    """Given a shape, return a list of render primitives in the form of [type, vertices, indices]
+    """
+    _check_shape(shape)
+    render_primitives = []
+    if isinstance(shape, Arc):
+        # Render meshes
+        if p5.renderer.fill_enabled:
+            render_primitives.extend(_get_meshes(shape))
+        # Render borders
+        if p5.renderer.stroke_enabled:
+            if shape.arc_mode in ['CHORD', 'OPEN']:  # Implies shape.shape_type == TESS
+                render_primitives.extend(_get_borders(shape))
+            elif shape.arc_mode is None:  # Implies shape.shape_type == TRIANGLE_FAN
+                render_primitives.append(_get_line_from_verts(shape.vertices[1:]))
+            elif shape.arc_mode == 'PIE':  # Implies shape.shape_type == TRIANGLE_FAN
+                render_primitives.append(_get_line_from_verts(shape.vertices))
+    else:
+        # Render points
+        if shape.shape_type == SType.POINTS:
+            render_primitives.append(_vertices_to_render_primitive(render_primitives, 'points'))
+        # Render meshes
+        if p5.renderer.fill_enabled:
+            render_primitives.extend(_get_meshes(shape))
+        # Render borders
+        if p5.renderer.stroke_enabled:
+            render_primitives.extend(_get_borders(shape))
+    return render_primitives
 
 
 # Abstract class that contains common code for OpenGL renderers
